Create Mon 10 May
Reference Number: 5
Date: Monday 10th May
Time: 8:21pm

Project Description:
Since the last git I have made one large change. Encryption. I have made it so that all passwords are no longer stored in plain text but instead in SHA256 Encrypted strings. This is all done within the code of the program so that if a database breach were to occur, the attackers would not be able to log into the user's accounts and make any important changes.

I made the Encryptions system a function so that I can use it easily for both Signup and Login Features. This means that the database never sees the users plain text password and that the code always stores the password as the SHA256 encrypted version.

I have also made some other minor changes such as the removal of old useless code. I am yet to add customisability to the database connection and set up system and am still working on increasing the speed of the updating function. I may make the code multithreaded meaning that i can run the updating of the code constantly as a background process.

Furthermore I still need to add tags and comments to my code to make it easier to interpret and understand.